Fix acceptance

+++ gavrilova.polina/S1/inputList.hpp
@@ -6,8 +6,8 @@
 
 namespace gavrilova
 {
-  using FLpairs = std::forward_list< std::pair< std::string, std::forward_list< int > > >;
-  std::forward_list< int > inputNumbers(std::istream& in, size_t& curLen);
+  using FLpairs = std::forward_list< std::pair< std::string, std::forward_list< unsigned long long > > >;
+  std::forward_list< unsigned long long > inputNumbers(std::istream& in, size_t& curLen);
 }
 
 #endif
